Add Municipalities API Endpoint
Based on the sample municipality API JSON which is in JSON-API format, from the front-end application we implemented the fast_jsonapi gem along with the necessary attributes for a municipality from the municipalities endpoint.

In order for the url_helper to work for the municipalities url we had to set the default host for each environment in its configuration file since it was challenging to access request.host from the serializer and model.